terraform/2-deploy-aisoc/orchestrator/function_app/requirements.txt
azure-functions
azure-identity
azure-ai-projects
requests
